+++ philly-cluster/workerdeploy.sh
@@ -47,8 +47,9 @@ function initialSetup()
 
     usermod -a -G systemd-journal $ADMIN_USERNAME
     usermod -a -G docker $ADMIN_USERNAME
-    mkdir /var/nfsshare
-    mkdir /var/nfs-mount
+    [[ ! -d /var/nfshare ]] && mkdir /var/nfsshare
+    [[ ! -d /var/nfs-mount ]] && mkdir /var/nfs-mount
+    [[ ! -d /var/gfs ]] && mkdir /var/gfs
 
     #if there is a datadisk mounted on sdc we partition it and format it
     if [[ -z $(fdisk -l /dev/sdc 2>&1 | grep "cannot open") ]];
@@ -60,12 +61,11 @@ function initialSetup()
         fi
     fi
 
-    ln -s /bin/mount /usr/bin/mount
-    ln -s /sbin/sysctl /usr/sbin/sysctl
-    ln -s /bin/bash /usr/bin/bash
-    ln -s /bin/true /usr/bin/true
-    ln -s /bin/mount /usr/bin/mount
-    ln -s /bin/chmod /usr/bin/chmod
+    [[ ! -f /usr/bin/mount ]] && ln -s /bin/mount /usr/bin/mount
+    [[ ! -f /usr/sbin/sysctl ]] && ln -s /sbin/sysctl /usr/sbin/sysctl
+    [[ ! -f /usr/bin/bash ]] && ln -s /bin/bash /usr/bin/bash
+    [[ ! -f /usr/bin/true ]] && ln -s /bin/true /usr/bin/true
+    [[ ! -f /usr/bin/chmod ]] && ln -s /bin/chmod /usr/bin/chmod
 
     echo "Initial setup done"
 }
@@ -73,9 +73,9 @@ function initialSetup()
 
 function applyCloudConfig()
 {
-    mkdir -p /var/lib/coreos-install
+    [[ ! -d /var/lib/coreos-install ]] && mkdir -p /var/lib/coreos-install
     curl "http://$LOAD_BALANCER_IP/cloud-config/$NAME.yml?reconfigure" -o /var/lib/coreos-install/user_data
-    [ -f "/var/lib/coreos-install/user_data" ] && coreos-cloudinit --from-file=/var/lib/coreos-install/user_data
+    [[ -f "/var/lib/coreos-install/user_data" ]] && coreos-cloudinit --from-file=/var/lib/coreos-install/user_data
 
     if [[ -z $(id -u core 2>&1 | grep "no such user") ]]; then
         echo "core ALL=(ALL) NOPASSWD: ALL" >> /etc/sudoers
@@ -84,9 +84,9 @@ function applyCloudConfig()
     sed -i "s/exit 0//g" /etc/rc.local
     {
         echo "LOAD_BALANCER_IP=$LOAD_BALANCER_IP"
-        echo '[ ! -f "/var/lib/coreos-install/user_data" ] &&'
+        echo '[[ ! -f "/var/lib/coreos-install/user_data" ]] &&'
         echo '    sudo curl "http://$LOAD_BALANCER_IP/cloud-config/$(hostname).yml?reconfigure" -o /var/lib/coreos-install/user_data'
-        echo '[ -f "/var/lib/coreos-install/user_data" ] &&'
+        echo '[[ -f "/var/lib/coreos-install/user_data" ]] &&'
         echo '    coreos-cloudinit --from-file=/var/lib/coreos-install/user_data'
     } >> /etc/rc.local
